add size_t, good path finding until the end

+++ solver/src/main.c
@@ -15,7 +15,7 @@
 
 int call_algo(char **map)
 {
-    int *pos = malloc(sizeof(int) * 2);
+    size_t *pos = malloc(sizeof(size_t) * 2);
     int come_from = 2;
     int ret_val = 0;
 
@@ -23,9 +23,9 @@ int call_algo(char **map)
     pos[1] = 0;
     ret_val = algo(map, pos, come_from);
     if (ret_val == -1)
-        map[my_strarraylen(map) - 1][(int)strlen(map[0]) - 1] = '*';
+        map[my_strarraylen(map) - 1][strlen(map[0]) - 1] = '*';
     else
-        map[my_strarraylen(map) - 1][(int)strlen(map[0]) - 1] = 'o';
+        map[my_strarraylen(map) - 1][strlen(map[0]) - 1] = 'o';
     for (int i = 0; map[i] != NULL; i++)
         printf("%s\n", map[i]);
     return (ret_val);
